requirements.txt
Flask==3.1.0
requests==2.32.3
python-dotenv==1.0.1
langchain-core==0.3.69
langchain-google-genai==2.0.11
langchain==0.3.26
langchain-community==0.3.27
chromadb==1.0.15
langchain-huggingface
PyMuPDF==1.25.3
unstructured==0.18.1
